requirements.txt
# requirements.txt
pyyaml
requests
dataclasses_json
moviepy
numpy
scipy
shotstack-sdk
jupyter
ruff
faker